Decouple store models into separate package
