+++ internal/app/updater.go
@@ -1,4 +1,4 @@
-package store
+package app
 
 import (
 	"context"
@@ -6,6 +6,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/leighmacdonald/gbans/internal/store"
 	"go.uber.org/zap"
 )
 
@@ -47,7 +48,7 @@ func (c *DataUpdater[T]) Start(ctx context.Context) {
 		select {
 		case <-c.updateChan:
 			newData, errUpdate := c.update()
-			if errUpdate != nil && !errors.Is(errUpdate, ErrNoResult) {
+			if errUpdate != nil && !errors.Is(errUpdate, store.ErrNoResult) {
 				c.log.Error("Failed to update data source", zap.Error(errUpdate))
 
 				return
