+++ internal/cmd/import.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/leighmacdonald/gbans/internal/config"
 	"github.com/leighmacdonald/gbans/internal/log"
+	"github.com/leighmacdonald/gbans/internal/model"
 	"github.com/leighmacdonald/gbans/internal/store"
 	"github.com/leighmacdonald/gbans/pkg/util"
 	"github.com/leighmacdonald/steamid/v3/steamid"
@@ -103,7 +104,7 @@ func importConnectionsCmd() *cobra.Command {
 
 				if _, playerFound := playerExistsCache[sid]; !playerFound {
 					// Satisfy fk
-					var person store.Person
+					var person model.Person
 					if errPerson := database.GetOrCreatePersonBySteamID(ctx, sid, &person); errPerson != nil {
 						rootLogger.Error("Failed to get person", zap.Error(errPerson))
 
@@ -148,7 +149,7 @@ func importConnectionsCmd() *cobra.Command {
 					continue
 				}
 
-				if errAdd := database.AddConnectionHistory(ctx, &store.PersonConnection{
+				if errAdd := database.AddConnectionHistory(ctx, &model.PersonConnection{
 					IPAddr:      ipAddr,
 					SteamID:     sid,
 					PersonaName: name,
@@ -236,7 +237,7 @@ func importMessagesCmd() *cobra.Command {
 
 				if _, playerFound := playerExistsCache[sid]; !playerFound {
 					// Satisfy fk
-					var person store.Person
+					var person model.Person
 					if errPerson := database.GetOrCreatePersonBySteamID(ctx, sid, &person); errPerson != nil {
 						rootLogger.Error("Failed to get person", zap.Error(errPerson))
 
@@ -253,7 +254,7 @@ func importMessagesCmd() *cobra.Command {
 
 				serverID, serverFound := serverCache[serverName]
 				if !serverFound {
-					var server store.Server
+					var server model.Server
 					if errServer := database.GetServerByName(ctx, serverName, &server, true, true); errServer != nil {
 						rootLogger.Error("Failed to get server", zap.Error(errServer))
 
@@ -298,7 +299,7 @@ func importMessagesCmd() *cobra.Command {
 					continue
 				}
 
-				if errAdd := database.AddChatHistory(ctx, &store.PersonMessage{
+				if errAdd := database.AddChatHistory(ctx, &model.PersonMessage{
 					SteamID:     sid,
 					PersonaName: name,
 					ServerID:    serverID,
