+++ internal/model/net.go
@@ -1,0 +1,17 @@
+package model
+
+import "net"
+
+type CIDRBlockSource struct {
+	CIDRBlockSourceID int    `json:"cidr_block_source_id"`
+	Name              string `json:"name"`
+	URL               string `json:"url"`
+	Enabled           bool   `json:"enabled"`
+	TimeStamped
+}
+
+type CIDRBlockWhitelist struct {
+	CIDRBlockWhitelistID int        `json:"cidr_block_whitelist_id"`
+	Address              *net.IPNet `json:"address"`
+	TimeStamped
+}
