+++ internal/model/news.go
@@ -1,0 +1,12 @@
+package model
+
+import "time"
+
+type NewsEntry struct {
+	NewsID      int       `json:"news_id"`
+	Title       string    `json:"title"`
+	BodyMD      string    `json:"body_md"`
+	IsPublished bool      `json:"is_published"`
+	CreatedOn   time.Time `json:"created_on,omitempty"`
+	UpdatedOn   time.Time `json:"updated_on,omitempty"`
+}
