+++ internal/store/filters.go
@@ -2,53 +2,14 @@ package store
 
 import (
 	"context"
-	"regexp"
-	"strings"
-	"time"
 
 	sq "github.com/Masterminds/squirrel"
+	"github.com/leighmacdonald/gbans/internal/model"
 	"github.com/leighmacdonald/steamid/v3/steamid"
 	"go.uber.org/zap"
 )
 
-type FilterAction int
-
-const (
-	Kick FilterAction = iota
-	Mute
-	Ban
-)
-
-type Filter struct {
-	FilterID     int64          `json:"filter_id"`
-	AuthorID     steamid.SID64  `json:"author_id"`
-	Pattern      string         `json:"pattern"`
-	IsRegex      bool           `json:"is_regex"`
-	IsEnabled    bool           `json:"is_enabled"`
-	Action       FilterAction   `json:"action"`
-	Duration     string         `json:"duration"`
-	Regex        *regexp.Regexp `json:"-"`
-	TriggerCount int64          `json:"trigger_count"`
-	Weight       int            `json:"weight"`
-	CreatedOn    time.Time      `json:"created_on"`
-	UpdatedOn    time.Time      `json:"updated_on"`
-}
-
-func (f *Filter) Init() {
-	if f.IsRegex {
-		f.Regex = regexp.MustCompile(f.Pattern)
-	}
-}
-
-func (f *Filter) Match(value string) bool {
-	if f.IsRegex {
-		return f.Regex.MatchString(strings.ToLower(value))
-	}
-
-	return f.Pattern == strings.ToLower(value)
-}
-
-func (db *Store) SaveFilter(ctx context.Context, filter *Filter) error {
+func (db *Store) SaveFilter(ctx context.Context, filter *model.Filter) error {
 	if filter.FilterID > 0 {
 		return db.updateFilter(ctx, filter)
 	} else {
@@ -56,7 +17,7 @@ func (db *Store) SaveFilter(ctx context.Context, filter *Filter) error {
 	}
 }
 
-func (db *Store) insertFilter(ctx context.Context, filter *Filter) error {
+func (db *Store) insertFilter(ctx context.Context, filter *model.Filter) error {
 	const query = `
 		INSERT INTO filtered_word (author_id, pattern, is_regex, is_enabled, trigger_count, created_on, updated_on, action, duration, weight) 
 		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10) 
@@ -73,7 +34,7 @@ func (db *Store) insertFilter(ctx context.Context, filter *Filter) error {
 	return nil
 }
 
-func (db *Store) updateFilter(ctx context.Context, filter *Filter) error {
+func (db *Store) updateFilter(ctx context.Context, filter *model.Filter) error {
 	query := db.sb.
 		Update("filtered_word").
 		Set("author_id", filter.AuthorID.Int64()).
@@ -97,7 +58,7 @@ func (db *Store) updateFilter(ctx context.Context, filter *Filter) error {
 	return nil
 }
 
-func (db *Store) DropFilter(ctx context.Context, filter *Filter) error {
+func (db *Store) DropFilter(ctx context.Context, filter *model.Filter) error {
 	query := db.sb.
 		Delete("filtered_word").
 		Where(sq.Eq{"filter_id": filter.FilterID})
@@ -112,7 +73,7 @@ func (db *Store) DropFilter(ctx context.Context, filter *Filter) error {
 	return nil
 }
 
-func (db *Store) GetFilterByID(ctx context.Context, filterID int64, filter *Filter) error {
+func (db *Store) GetFilterByID(ctx context.Context, filterID int64, filter *model.Filter) error {
 	query := db.sb.
 		Select("filter_id", "author_id", "pattern", "is_regex",
 			"is_enabled", "trigger_count", "created_on", "updated_on", "action", "duration", "weight").
@@ -145,7 +106,7 @@ type FiltersQueryFilter struct {
 	QueryFilter
 }
 
-func (db *Store) GetFilters(ctx context.Context, opts FiltersQueryFilter) ([]Filter, int64, error) {
+func (db *Store) GetFilters(ctx context.Context, opts FiltersQueryFilter) ([]model.Filter, int64, error) {
 	builder := db.sb.
 		Select("f.filter_id", "f.author_id", "f.pattern", "f.is_regex",
 			"f.is_enabled", "f.trigger_count", "f.created_on", "f.updated_on", "f.action", "f.duration", "f.weight").
@@ -167,11 +128,11 @@ func (db *Store) GetFilters(ctx context.Context, opts FiltersQueryFilter) ([]Fil
 
 	defer rows.Close()
 
-	var filters []Filter
+	var filters []model.Filter
 
 	for rows.Next() {
 		var (
-			filter   Filter
+			filter   model.Filter
 			authorID int64
 		)
