+++ internal/store/match.go
@@ -2,15 +2,12 @@ package store
 
 import (
 	"context"
-	"fmt"
-	"math"
-	"sort"
 	"time"
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/gofrs/uuid/v5"
 	"github.com/jackc/pgx/v5"
-	"github.com/leighmacdonald/gbans/pkg/fp"
+	"github.com/leighmacdonald/gbans/internal/model"
 	"github.com/leighmacdonald/gbans/pkg/logparse"
 	"github.com/leighmacdonald/steamid/v3/steamid"
 	"github.com/pkg/errors"
@@ -26,168 +23,7 @@ type MatchesQueryOpts struct {
 	TimeEnd   *time.Time    `json:"time_end,omitempty"`
 }
 
-type MatchPlayerKillstreak struct {
-	MatchKillstreakID int64                `json:"match_killstreak_id"`
-	MatchPlayerID     int64                `json:"match_player_id"`
-	PlayerClass       logparse.PlayerClass `json:"player_class"`
-	Killstreak        int                  `json:"killstreak"`
-	// Seconds
-	Duration int `json:"duration"`
-}
-
-type MatchPlayerClass struct {
-	MatchPlayerClassID int                  `json:"match_player_class_id"`
-	MatchPlayerID      int64                `json:"match_player_id"`
-	PlayerClass        logparse.PlayerClass `json:"player_class"`
-	Kills              int                  `json:"kills"`
-	Assists            int                  `json:"assists"`
-	Deaths             int                  `json:"deaths"`
-	Playtime           int                  `json:"playtime"`
-	Dominations        int                  `json:"dominations"`
-	Dominated          int                  `json:"dominated"`
-	Revenges           int                  `json:"revenges"`
-	Damage             int                  `json:"damage"`
-	DamageTaken        int                  `json:"damage_taken"`
-	HealingTaken       int                  `json:"healing_taken"`
-	Captures           int                  `json:"captures"`
-	CapturesBlocked    int                  `json:"captures_blocked"`
-	BuildingDestroyed  int                  `json:"building_destroyed"`
-}
-
-type MatchPlayer struct {
-	MatchPlayerID int64 `json:"match_player_id"`
-	CommonPlayerStats
-	Team      logparse.Team `json:"team"`
-	TimeStart time.Time     `json:"time_start"`
-	TimeEnd   time.Time     `json:"time_end"`
-
-	MedicStats  *MatchHealer            `json:"medic_stats"`
-	Classes     []MatchPlayerClass      `json:"classes"`
-	Killstreaks []MatchPlayerKillstreak `json:"killstreaks"`
-	Weapons     []MatchPlayerWeapon     `json:"weapons"`
-}
-
-func (player MatchPlayer) BiggestKillstreak() *MatchPlayerKillstreak {
-	var biggest *MatchPlayerKillstreak
-
-	for _, killstreakVal := range player.Killstreaks {
-		killstreak := killstreakVal
-		if biggest == nil || killstreak.Killstreak > biggest.Killstreak {
-			biggest = &killstreak
-		}
-	}
-
-	return biggest
-}
-
-func (player MatchPlayer) KDRatio() float64 {
-	if player.Deaths <= 0 {
-		return -1
-	}
-
-	return math.Ceil((float64(player.Kills)/float64(player.Deaths))*100) / 100
-}
-
-func (player MatchPlayer) KDARatio() float64 {
-	if player.Deaths <= 0 {
-		return -1
-	}
-
-	return math.Ceil((float64(player.Kills+player.Assists)/float64(player.Deaths))*100) / 100
-}
-
-func (player MatchPlayer) DamagePerMin() int {
-	return fp.Max[int](int(float64(player.Damage)/player.TimeEnd.Sub(player.TimeStart).Minutes()), 0)
-}
-
-type MatchHealer struct {
-	MatchMedicID        int64   `json:"match_medic_id"`
-	MatchPlayerID       int64   `json:"match_player_id"`
-	Healing             int     `json:"healing"`
-	ChargesUber         int     `json:"charges_uber"`
-	ChargesKritz        int     `json:"charges_kritz"`
-	ChargesVacc         int     `json:"charges_vacc"`
-	ChargesQuickfix     int     `json:"charges_quickfix"`
-	Drops               int     `json:"drops"`
-	NearFullChargeDeath int     `json:"near_full_charge_death"`
-	AvgUberLength       float32 `json:"avg_uber_length"`
-	MajorAdvLost        int     `json:"major_adv_lost"`
-	BiggestAdvLost      int     `json:"biggest_adv_lost"`
-}
-
-func (h MatchHealer) HealingPerMin(matchDuration time.Duration) int {
-	if h.Healing <= 0 {
-		return 0
-	}
-
-	return int(float64(h.Healing) / matchDuration.Minutes())
-}
-
-type MatchWeapon struct {
-	PlayerWeaponID int64 `json:"player_weapon_id"`
-	MatchPlayerID  int64 `json:"match_player_id"`
-}
-
-type MatchResult struct {
-	MatchID    uuid.UUID           `json:"match_id"`
-	ServerID   int                 `json:"server_id"`
-	Title      string              `json:"title"`
-	MapName    string              `json:"map_name"`
-	TeamScores logparse.TeamScores `json:"team_scores"`
-	TimeStart  time.Time           `json:"time_start"`
-	TimeEnd    time.Time           `json:"time_end"`
-	Winner     logparse.Team       `json:"winner"`
-	Players    []*MatchPlayer      `json:"players"`
-	Chat       PersonMessages      `json:"chat"`
-}
-
-func (match *MatchResult) TopPlayers() []*MatchPlayer {
-	players := match.Players
-
-	sort.SliceStable(players, func(i, j int) bool {
-		return players[i].Kills > players[j].Kills
-	})
-
-	return players
-}
-
-func (match *MatchResult) TopKillstreaks(count int) []*MatchPlayer {
-	var killStreakPlayers []*MatchPlayer
-
-	for _, player := range match.Players {
-		if killStreak := player.BiggestKillstreak(); killStreak != nil {
-			killStreakPlayers = append(killStreakPlayers, player)
-		}
-	}
-
-	sort.SliceStable(killStreakPlayers, func(i, j int) bool {
-		return killStreakPlayers[i].BiggestKillstreak().Killstreak > killStreakPlayers[j].BiggestKillstreak().Killstreak
-	})
-
-	if len(killStreakPlayers) > count {
-		return killStreakPlayers[0:count]
-	}
-
-	return killStreakPlayers
-}
-
-func (match *MatchResult) Healers() []*MatchPlayer {
-	var healers []*MatchPlayer
-
-	for _, player := range match.Players {
-		if player.MedicStats != nil {
-			healers = append(healers, player)
-		}
-	}
-
-	sort.SliceStable(healers, func(i, j int) bool {
-		return healers[i].MedicStats.Healing > healers[j].MedicStats.Healing
-	})
-
-	return healers
-}
-
-func (db *Store) matchGetPlayerClasses(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64][]MatchPlayerClass, error) {
+func (db *Store) matchGetPlayerClasses(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64][]model.MatchPlayerClass, error) {
 	const query = `
 		SELECT mp.steam_id, c.match_player_class_id, c.match_player_id, c.player_class_id, c.kills, 
 		   c.assists, c.deaths, c.playtime, c.dominations, c.dominated, c.revenges, c.damage, c.damage_taken, c.healing_taken,
@@ -203,12 +39,12 @@ func (db *Store) matchGetPlayerClasses(ctx context.Context, matchID uuid.UUID) (
 
 	defer rows.Close()
 
-	results := map[steamid.SID64][]MatchPlayerClass{}
+	results := map[steamid.SID64][]model.MatchPlayerClass{}
 
 	for rows.Next() {
 		var (
 			steamID int64
-			stats   MatchPlayerClass
+			stats   model.MatchPlayerClass
 		)
 
 		if errScan := rows.
@@ -223,7 +59,7 @@ func (db *Store) matchGetPlayerClasses(ctx context.Context, matchID uuid.UUID) (
 
 		res, found := results[sid]
 		if !found {
-			res = []MatchPlayerClass{}
+			res = []model.MatchPlayerClass{}
 		}
 
 		results[sid] = append(res, stats)
@@ -236,19 +72,7 @@ func (db *Store) matchGetPlayerClasses(ctx context.Context, matchID uuid.UUID) (
 	return results, nil
 }
 
-type MatchPlayerWeapon struct {
-	Weapon
-	Kills     int     `json:"kills"`
-	Damage    int     `json:"damage"`
-	Shots     int     `json:"shots"`
-	Hits      int     `json:"hits"`
-	Accuracy  float64 `json:"accuracy"`
-	Backstabs int     `json:"backstabs"`
-	Headshots int     `json:"headshots"`
-	Airshots  int     `json:"airshots"`
-}
-
-func (db *Store) matchGetPlayerWeapons(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64][]MatchPlayerWeapon, error) {
+func (db *Store) matchGetPlayerWeapons(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64][]model.MatchPlayerWeapon, error) {
 	const query = `
 		SELECT mp.steam_id, mw.weapon_id, w.name, w.key,  mw.kills, mw.damage, mw.shots, mw.hits, mw.backstabs, mw.headshots, mw.airshots
 		FROM match m
@@ -258,7 +82,7 @@ func (db *Store) matchGetPlayerWeapons(ctx context.Context, matchID uuid.UUID) (
 		WHERE m.match_id = $1 and mw.weapon_id is not null
 		ORDER BY mw.kills DESC`
 
-	results := map[steamid.SID64][]MatchPlayerWeapon{}
+	results := map[steamid.SID64][]model.MatchPlayerWeapon{}
 
 	rows, errRows := db.Query(ctx, query, matchID)
 	if errRows != nil {
@@ -270,7 +94,7 @@ func (db *Store) matchGetPlayerWeapons(ctx context.Context, matchID uuid.UUID) (
 	for rows.Next() {
 		var (
 			steamID int64
-			mpw     MatchPlayerWeapon
+			mpw     model.MatchPlayerWeapon
 		)
 
 		if errScan := rows.
@@ -283,7 +107,7 @@ func (db *Store) matchGetPlayerWeapons(ctx context.Context, matchID uuid.UUID) (
 
 		res, found := results[sid]
 		if !found {
-			res = []MatchPlayerWeapon{}
+			res = []model.MatchPlayerWeapon{}
 		}
 
 		results[sid] = append(res, mpw)
@@ -292,7 +116,7 @@ func (db *Store) matchGetPlayerWeapons(ctx context.Context, matchID uuid.UUID) (
 	return results, nil
 }
 
-func (db *Store) matchGetPlayerKillstreak(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64][]MatchPlayerKillstreak, error) {
+func (db *Store) matchGetPlayerKillstreak(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64][]model.MatchPlayerKillstreak, error) {
 	const query = `
 		SELECT mp.steam_id, k.match_player_id, k.player_class_id, k.killstreak, k.duration
 		FROM match_player_killstreak k
@@ -306,12 +130,12 @@ func (db *Store) matchGetPlayerKillstreak(ctx context.Context, matchID uuid.UUID
 
 	defer rows.Close()
 
-	results := map[steamid.SID64][]MatchPlayerKillstreak{}
+	results := map[steamid.SID64][]model.MatchPlayerKillstreak{}
 
 	for rows.Next() {
 		var (
 			steamID int64
-			stats   MatchPlayerKillstreak
+			stats   model.MatchPlayerKillstreak
 		)
 
 		if errScan := rows.
@@ -323,7 +147,7 @@ func (db *Store) matchGetPlayerKillstreak(ctx context.Context, matchID uuid.UUID
 
 		res, found := results[sid]
 		if !found {
-			res = []MatchPlayerKillstreak{}
+			res = []model.MatchPlayerKillstreak{}
 		}
 
 		results[sid] = append(res, stats)
@@ -332,7 +156,7 @@ func (db *Store) matchGetPlayerKillstreak(ctx context.Context, matchID uuid.UUID
 	return results, nil
 }
 
-func (db *Store) matchGetPlayers(ctx context.Context, matchID uuid.UUID) ([]*MatchPlayer, error) {
+func (db *Store) matchGetPlayers(ctx context.Context, matchID uuid.UUID) ([]*model.MatchPlayer, error) {
 	const queryPlayer = `
 		SELECT
 			p.match_player_id,
@@ -384,12 +208,12 @@ func (db *Store) matchGetPlayers(ctx context.Context, matchID uuid.UUID) ([]*Mat
 			p.match_player_id, w.kills, w.damage, w.shots, w.hits, w.backstabs, w.headshots, w.airshots, pe.steam_id
 		ORDER BY w.kills DESC`
 
-	var players []*MatchPlayer
+	var players []*model.MatchPlayer
 
 	playerRows, errPlayer := db.Query(ctx, queryPlayer, matchID)
 	if errPlayer != nil {
 		if errors.Is(errPlayer, ErrNoResult) {
-			return []*MatchPlayer{}, nil
+			return []*model.MatchPlayer{}, nil
 		}
 
 		return nil, errors.Wrapf(errPlayer, "Failed to query match players")
@@ -399,7 +223,7 @@ func (db *Store) matchGetPlayers(ctx context.Context, matchID uuid.UUID) ([]*Mat
 
 	for playerRows.Next() {
 		var (
-			mpSum   = &MatchPlayer{}
+			mpSum   = &model.MatchPlayer{}
 			steamID int64
 		)
 
@@ -420,7 +244,7 @@ func (db *Store) matchGetPlayers(ctx context.Context, matchID uuid.UUID) ([]*Mat
 	return players, nil
 }
 
-func (db *Store) matchGetMedics(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64]MatchHealer, error) {
+func (db *Store) matchGetMedics(ctx context.Context, matchID uuid.UUID) (map[steamid.SID64]model.MatchHealer, error) {
 	const query = `
 		SELECT m.match_medic_id,
 			   mp.steam_id,
@@ -438,7 +262,7 @@ func (db *Store) matchGetMedics(ctx context.Context, matchID uuid.UUID) (map[ste
 		LEFT JOIN match_player mp on mp.match_player_id = m.match_player_id
 		WHERE mp.match_id = $1`
 
-	medics := map[steamid.SID64]MatchHealer{}
+	medics := map[steamid.SID64]model.MatchHealer{}
 
 	medicRows, errMedics := db.Query(ctx, query, matchID)
 	if errMedics != nil {
@@ -453,7 +277,7 @@ func (db *Store) matchGetMedics(ctx context.Context, matchID uuid.UUID) (map[ste
 
 	for medicRows.Next() {
 		var (
-			mpSum   = MatchHealer{}
+			mpSum   = model.MatchHealer{}
 			steamID int64
 		)
 
@@ -477,7 +301,7 @@ func (db *Store) matchGetMedics(ctx context.Context, matchID uuid.UUID) (map[ste
 	return medics, nil
 }
 
-func (db *Store) matchGetChat(ctx context.Context, matchID uuid.UUID) (PersonMessages, error) {
+func (db *Store) matchGetChat(ctx context.Context, matchID uuid.UUID) (model.PersonMessages, error) {
 	const query = `
 		SELECT c.person_message_id, c.steam_id, c.server_id, c.body, c.persona_name, c.team, 
 		       c.created_on, c.match_id, COUNT(f.person_message_id)::int::boolean as flagged
@@ -487,7 +311,7 @@ func (db *Store) matchGetChat(ctx context.Context, matchID uuid.UUID) (PersonMes
 		GROUP BY c.person_message_id
 		`
 
-	messages := PersonMessages{}
+	messages := model.PersonMessages{}
 
 	medicRows, errMedics := db.Query(ctx, query, matchID)
 	if errMedics != nil {
@@ -502,7 +326,7 @@ func (db *Store) matchGetChat(ctx context.Context, matchID uuid.UUID) (PersonMes
 
 	for medicRows.Next() {
 		var (
-			msg     PersonMessage
+			msg     model.PersonMessage
 			steamID int64
 		)
 
@@ -524,7 +348,7 @@ func (db *Store) matchGetChat(ctx context.Context, matchID uuid.UUID) (PersonMes
 	return messages, nil
 }
 
-func (db *Store) MatchGetByID(ctx context.Context, matchID uuid.UUID, match *MatchResult) error {
+func (db *Store) MatchGetByID(ctx context.Context, matchID uuid.UUID, match *model.MatchResult) error {
 	const query = `
 		SELECT match_id, server_id, map, title, score_red, score_blu, time_red, time_blu, time_start, time_end, winner
 		FROM match WHERE match_id = $1`
@@ -609,20 +433,20 @@ func (db *Store) MatchGetByID(ctx context.Context, matchID uuid.UUID, match *Mat
 	match.Chat = chat
 
 	if match.Chat == nil {
-		match.Chat = PersonMessages{}
+		match.Chat = model.PersonMessages{}
 	}
 
 	for _, player := range match.Players {
 		if player.Weapons == nil {
-			player.Weapons = []MatchPlayerWeapon{}
+			player.Weapons = []model.MatchPlayerWeapon{}
 		}
 
 		if player.Classes == nil {
-			player.Classes = []MatchPlayerClass{}
+			player.Classes = []model.MatchPlayerClass{}
 		}
 
 		if player.Killstreaks == nil {
-			player.Killstreaks = []MatchPlayerKillstreak{}
+			player.Killstreaks = []model.MatchPlayerKillstreak{}
 		}
 	}
 
@@ -676,7 +500,7 @@ func (db *Store) MatchSave(ctx context.Context, match *logparse.Match) error {
 			continue
 		}
 
-		var loadPlayerTest Person
+		var loadPlayerTest model.Person
 		if errPlayer := db.GetOrCreatePersonBySteamID(ctx, player.SteamID, &loadPlayerTest); errPlayer != nil {
 			if errRollback := transaction.Rollback(ctx); errRollback != nil {
 				db.log.Error("Failed to rollback tx", zap.Error(errRollback))
@@ -838,151 +662,7 @@ func (db *Store) saveMatchKillstreakStats(ctx context.Context, transaction pgx.T
 	return nil
 }
 
-type PlayerClassStats struct {
-	Class              logparse.PlayerClass
-	ClassName          string
-	Kills              int
-	Assists            int
-	Deaths             int
-	Damage             int
-	Dominations        int
-	Dominated          int
-	Revenges           int
-	DamageTaken        int
-	HealingTaken       int
-	HealthPacks        int
-	Captures           int
-	CapturesBlocked    int
-	Extinguishes       int
-	BuildingsBuilt     int
-	BuildingsDestroyed int
-	Playtime           float64 // seconds
-}
-
-func (player PlayerClassStats) KDRatio() float64 {
-	if player.Deaths <= 0 {
-		return -1
-	}
-
-	return math.Ceil((float64(player.Kills)/float64(player.Deaths))*100) / 100
-}
-
-func (player PlayerClassStats) KDARatio() float64 {
-	if player.Deaths <= 0 {
-		return -1
-	}
-
-	return math.Ceil((float64(player.Kills+player.Assists)/float64(player.Deaths))*100) / 100
-}
-
-func (player PlayerClassStats) DamagePerMin() int {
-	return int(float64(player.Damage) / (player.Playtime / 60))
-}
-
-type PlayerClassStatsCollection []PlayerClassStats
-
-func (ps PlayerClassStatsCollection) Kills() int {
-	var total int
-	for _, class := range ps {
-		total += class.Kills
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Assists() int {
-	var total int
-	for _, class := range ps {
-		total += class.Assists
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Deaths() int {
-	var total int
-	for _, class := range ps {
-		total += class.Deaths
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Damage() int {
-	var total int
-	for _, class := range ps {
-		total += class.Damage
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) DamageTaken() int {
-	var total int
-	for _, class := range ps {
-		total += class.DamageTaken
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Captures() int {
-	var total int
-	for _, class := range ps {
-		total += class.Captures
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Dominations() int {
-	var total int
-	for _, class := range ps {
-		total += class.Dominations
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Dominated() int {
-	var total int
-	for _, class := range ps {
-		total += class.Dominated
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) Playtime() float64 {
-	var total float64
-	for _, class := range ps {
-		total += class.Playtime
-	}
-
-	return total
-}
-
-func (ps PlayerClassStatsCollection) DamagePerMin() int {
-	return int(float64(ps.Damage()) / (ps.Playtime() / 60))
-}
-
-func (ps PlayerClassStatsCollection) KDRatio() float64 {
-	if ps.Deaths() <= 0 {
-		return -1
-	}
-
-	return math.Ceil((float64(ps.Kills())/float64(ps.Deaths()))*100) / 100
-}
-
-func (ps PlayerClassStatsCollection) KDARatio() float64 {
-	if ps.Deaths() <= 0 {
-		return -1
-	}
-
-	return math.Ceil((float64(ps.Kills()+ps.Assists())/float64(ps.Deaths()))*100) / 100
-}
-
-func (db *Store) StatsPlayerClass(ctx context.Context, sid64 steamid.SID64) (PlayerClassStatsCollection, error) {
+func (db *Store) StatsPlayerClass(ctx context.Context, sid64 steamid.SID64) (model.PlayerClassStatsCollection, error) {
 	const query = `
 		SELECT c.player_class_id,
 			   coalesce(SUM(c.kills), 0)               as kill,
@@ -1014,10 +694,10 @@ func (db *Store) StatsPlayerClass(ctx context.Context, sid64 steamid.SID64) (Pla
 
 	defer rows.Close()
 
-	var stats PlayerClassStatsCollection
+	var stats model.PlayerClassStatsCollection
 
 	for rows.Next() {
-		var class PlayerClassStats
+		var class model.PlayerClassStats
 		if errScan := rows.
 			Scan(&class.Class, &class.Kills, &class.Damage, &class.Assists, &class.Deaths, &class.Dominations,
 				&class.Dominated, &class.Revenges, &class.DamageTaken, &class.HealingTaken, &class.HealthPacks,
@@ -1033,27 +713,7 @@ func (db *Store) StatsPlayerClass(ctx context.Context, sid64 steamid.SID64) (Pla
 	return stats, nil
 }
 
-type PlayerWeaponStats struct {
-	Weapon     logparse.Weapon
-	WeaponName string
-	Kills      int
-	Damage     int
-	Shots      int
-	Hits       int
-	Backstabs  int
-	Headshots  int
-	Airshots   int
-}
-
-func (ws PlayerWeaponStats) Accuracy() float64 {
-	if ws.Shots == 0 {
-		return 0
-	}
-
-	return math.Ceil(float64(ws.Hits)/float64(ws.Shots)*10000) / 100
-}
-
-func (db *Store) StatsPlayerWeapons(ctx context.Context, sid64 steamid.SID64) ([]PlayerWeaponStats, error) {
+func (db *Store) StatsPlayerWeapons(ctx context.Context, sid64 steamid.SID64) ([]model.PlayerWeaponStats, error) {
 	const query = `
 		SELECT n.key,
 			   n.name,
@@ -1078,10 +738,10 @@ func (db *Store) StatsPlayerWeapons(ctx context.Context, sid64 steamid.SID64) ([
 
 	defer rows.Close()
 
-	var stats []PlayerWeaponStats
+	var stats []model.PlayerWeaponStats
 
 	for rows.Next() {
-		var class PlayerWeaponStats
+		var class model.PlayerWeaponStats
 		if errScan := rows.
 			Scan(&class.Weapon, &class.WeaponName, &class.Kills, &class.Damage, &class.Shots, &class.Hits,
 				&class.Backstabs, &class.Headshots, &class.Airshots); errScan != nil {
@@ -1094,15 +754,7 @@ func (db *Store) StatsPlayerWeapons(ctx context.Context, sid64 steamid.SID64) ([
 	return stats, nil
 }
 
-type PlayerKillstreakStats struct {
-	Class     logparse.PlayerClass `json:"class"`
-	ClassName string               `json:"class_name"`
-	Kills     int                  `json:"kills"`
-	Duration  int                  `json:"duration"`
-	CreatedOn time.Time            `json:"created_on"`
-}
-
-func (db *Store) StatsPlayerKillstreaks(ctx context.Context, sid64 steamid.SID64) ([]PlayerKillstreakStats, error) {
+func (db *Store) StatsPlayerKillstreaks(ctx context.Context, sid64 steamid.SID64) ([]model.PlayerKillstreakStats, error) {
 	const query = `
 		SELECT k.player_class_id,
 			   SUM(k.killstreak) as killstreak,
@@ -1125,10 +777,10 @@ func (db *Store) StatsPlayerKillstreaks(ctx context.Context, sid64 steamid.SID64
 
 	defer rows.Close()
 
-	var stats []PlayerKillstreakStats
+	var stats []model.PlayerKillstreakStats
 
 	for rows.Next() {
-		var class PlayerKillstreakStats
+		var class model.PlayerKillstreakStats
 		if errScan := rows.
 			Scan(&class.Class, &class.Kills, &class.Duration, &class.CreatedOn); errScan != nil {
 			return nil, Err(errScan)
@@ -1141,18 +793,7 @@ func (db *Store) StatsPlayerKillstreaks(ctx context.Context, sid64 steamid.SID64
 	return stats, nil
 }
 
-type PlayerMedicStats struct {
-	Healing             int
-	Drops               int
-	NearFullChargeDeath int
-	AvgUberLength       float64
-	ChargesUber         int
-	ChargesKritz        int
-	ChargesVacc         int
-	ChargesQuickfix     int
-}
-
-func (db *Store) StatsPlayerMedic(ctx context.Context, sid64 steamid.SID64) ([]PlayerMedicStats, error) {
+func (db *Store) StatsPlayerMedic(ctx context.Context, sid64 steamid.SID64) ([]model.PlayerMedicStats, error) {
 	const query = `
 		SELECT coalesce(SUM(m.healing), 0)                as healing,
 			   coalesce(SUM(m.drops), 0)                  as drops,
@@ -1174,10 +815,10 @@ func (db *Store) StatsPlayerMedic(ctx context.Context, sid64 steamid.SID64) ([]P
 
 	defer rows.Close()
 
-	var stats []PlayerMedicStats
+	var stats []model.PlayerMedicStats
 
 	for rows.Next() {
-		var class PlayerMedicStats
+		var class model.PlayerMedicStats
 		if errScan := rows.
 			Scan(&class.Healing, &class.Drops, &class.NearFullChargeDeath, &class.AvgUberLength,
 				&class.ChargesUber, &class.ChargesKritz, &class.ChargesVacc, &class.ChargesQuickfix); errScan != nil {
@@ -1190,42 +831,7 @@ func (db *Store) StatsPlayerMedic(ctx context.Context, sid64 steamid.SID64) ([]P
 	return stats, nil
 }
 
-type CommonPlayerStats struct {
-	SteamID           steamid.SID64 `json:"steam_id"`
-	Name              string        `json:"name"`
-	AvatarHash        string        `json:"avatar_hash"`
-	Kills             int           `json:"kills"`
-	Assists           int           `json:"assists"`
-	Deaths            int           `json:"deaths"`
-	Suicides          int           `json:"suicides"`
-	Dominations       int           `json:"dominations"`
-	Dominated         int           `json:"dominated"`
-	Revenges          int           `json:"revenges"`
-	Damage            int           `json:"damage"`
-	DamageTaken       int           `json:"damage_taken"`
-	HealingTaken      int           `json:"healing_taken"`
-	HealthPacks       int           `json:"health_packs"`
-	HealingPacks      int           `json:"healing_packs"` // Healing from packs
-	Captures          int           `json:"captures"`
-	CapturesBlocked   int           `json:"captures_blocked"`
-	Extinguishes      int           `json:"extinguishes"`
-	BuildingBuilt     int           `json:"building_built"`
-	BuildingDestroyed int           `json:"building_destroyed"` // Opposing team buildings
-	Backstabs         int           `json:"backstabs"`
-	Airshots          int           `json:"airshots"`
-	Headshots         int           `json:"headshots"`
-	Shots             int           `json:"shots"`
-	Hits              int           `json:"hits"`
-}
-type PlayerStats struct {
-	CommonPlayerStats
-	PlayerMedicStats
-	MatchesTotal int           `json:"matches_total"`
-	MatchesWon   int           `json:"matches_won"`
-	PlayTime     time.Duration `json:"play_time"`
-}
-
-func (db *Store) PlayerStats(ctx context.Context, steamID steamid.SID64, stats *PlayerStats) error {
+func (db *Store) PlayerStats(ctx context.Context, steamID steamid.SID64, stats *model.PlayerStats) error {
 	const query = `
 		SELECT count(m.match_id)            as                     matches,
 			   sum(case when mp.team = m.winner then 1 else 0 end) wins,
@@ -1275,24 +881,7 @@ func (db *Store) PlayerStats(ctx context.Context, steamID steamid.SID64, stats *
 	return nil
 }
 
-type MatchSummary struct {
-	MatchID   uuid.UUID `json:"match_id"`
-	ServerID  int       `json:"server_id"`
-	IsWinner  bool      `json:"is_winner"`
-	ShortName string    `json:"short_name"`
-	Title     string    `json:"title"`
-	MapName   string    `json:"map_name"`
-	ScoreBlu  int       `json:"score_blu"`
-	ScoreRed  int       `json:"score_red"`
-	TimeStart time.Time `json:"time_start"`
-	TimeEnd   time.Time `json:"time_end"`
-}
-
-func (m MatchSummary) Path() string {
-	return fmt.Sprintf("/log/%s", m.MatchID.String())
-}
-
-func (db *Store) Matches(ctx context.Context, opts MatchesQueryOpts) ([]MatchSummary, int64, error) {
+func (db *Store) Matches(ctx context.Context, opts MatchesQueryOpts) ([]model.MatchSummary, int64, error) {
 	countBuilder := db.sb.Select("count(m.match_id) as count").
 		From("match m").
 		LeftJoin("public.match_player mp on m.match_id = mp.match_id").
@@ -1351,10 +940,10 @@ func (db *Store) Matches(ctx context.Context, opts MatchesQueryOpts) ([]MatchSum
 
 	defer rows.Close()
 
-	var matches []MatchSummary
+	var matches []model.MatchSummary
 
 	for rows.Next() {
-		var summary MatchSummary
+		var summary model.MatchSummary
 		if errScan := rows.Scan(&summary.MatchID, &summary.ServerID, &summary.IsWinner, &summary.ShortName,
 			&summary.Title, &summary.MapName, &summary.ScoreBlu, &summary.ScoreRed, &summary.TimeStart,
 			&summary.TimeEnd); errScan != nil {
