+++ internal/store/server.go
@@ -2,91 +2,16 @@ package store
 
 import (
 	"context"
-	"fmt"
-	"net"
 	"time"
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/leighmacdonald/gbans/internal/consts"
+	"github.com/leighmacdonald/gbans/internal/model"
 	"github.com/leighmacdonald/steamid/v3/steamid"
 	"github.com/pkg/errors"
 )
 
-type ServerPermission struct {
-	SteamID         steamid.SID      `json:"steam_id"`
-	PermissionLevel consts.Privilege `json:"permission_level"`
-	Flags           string           `json:"flags"`
-}
-
-func NewServer(shortName string, address string, port int) Server {
-	return Server{
-		ShortName:      shortName,
-		Address:        address,
-		Port:           port,
-		RCON:           SecureRandomString(10),
-		ReservedSlots:  0,
-		Password:       SecureRandomString(10),
-		IsEnabled:      true,
-		EnableStats:    true,
-		TokenCreatedOn: time.Unix(0, 0),
-		CreatedOn:      time.Now(),
-		UpdatedOn:      time.Now(),
-	}
-}
-
-type Server struct {
-	// Auto generated id
-	ServerID int `db:"server_id" json:"server_id"`
-	// ShortName is a short reference name for the server eg: us-1
-	ShortName string `json:"short_name"`
-	Name      string `json:"name"`
-	// Address is the ip of the server
-	Address string `db:"address" json:"address"`
-	// Port is the port of the server
-	Port int `db:"port" json:"port"`
-	// RCON is the RCON password for the server
-	RCON          string `db:"rcon" json:"rcon"`
-	ReservedSlots int    `db:"reserved_slots" json:"reserved_slots"`
-	// Password is what the server uses to generate a token to make authenticated calls (permanent refresh token)
-	Password    string  `db:"password" json:"password"`
-	IsEnabled   bool    `json:"is_enabled"`
-	Deleted     bool    `json:"deleted"`
-	Region      string  `json:"region"`
-	CC          string  `json:"cc"`
-	Latitude    float64 `json:"latitude"`
-	Longitude   float64 `json:"longitude"`
-	LogSecret   int     `json:"log_secret"`
-	EnableStats bool    `json:"enable_stats"`
-	// TokenCreatedOn is set when changing the token
-	TokenCreatedOn time.Time `db:"token_created_on" json:"token_created_on"`
-	CreatedOn      time.Time `db:"created_on" json:"created_on"`
-	UpdatedOn      time.Time `db:"updated_on" json:"updated_on"`
-}
-
-func (s Server) IP(ctx context.Context) (net.IP, error) {
-	parsedIP := net.ParseIP(s.Address)
-	if parsedIP != nil {
-		// We already have an ip
-		return parsedIP, nil
-	}
-	// TODO proper timeout for ctx
-	ips, errResolve := net.DefaultResolver.LookupIP(ctx, "ip4", s.Address)
-	if errResolve != nil || len(ips) == 0 {
-		return nil, errors.Wrap(errResolve, "Could not resolve address")
-	}
-
-	return ips[0], nil
-}
-
-func (s Server) Addr() string {
-	return fmt.Sprintf("%s:%d", s.Address, s.Port)
-}
-
-func (s Server) Slots(statusSlots int) int {
-	return statusSlots - s.ReservedSlots
-}
-
-func (db *Store) GetServer(ctx context.Context, serverID int, server *Server) error {
+func (db *Store) GetServer(ctx context.Context, serverID int, server *model.Server) error {
 	row, rowErr := db.QueryRowBuilder(ctx, db.sb.
 		Select("server_id", "short_name", "name", "address", "port", "rcon", "password",
 			"token_created_on", "created_on", "updated_on", "reserved_slots", "is_enabled", "region", "cc",
@@ -108,7 +33,7 @@ func (db *Store) GetServer(ctx context.Context, serverID int, server *Server) er
 	return nil
 }
 
-func (db *Store) GetServerPermissions(ctx context.Context) ([]ServerPermission, error) {
+func (db *Store) GetServerPermissions(ctx context.Context) ([]model.ServerPermission, error) {
 	rows, errRows := db.QueryBuilder(ctx, db.sb.
 		Select("steam_id", "permission_level").From("person").
 		Where(sq.GtOrEq{"permission_level": consts.PReserved}).
@@ -119,7 +44,7 @@ func (db *Store) GetServerPermissions(ctx context.Context) ([]ServerPermission, 
 
 	defer rows.Close()
 
-	var perms []ServerPermission
+	var perms []model.ServerPermission
 
 	for rows.Next() {
 		var (
@@ -143,7 +68,7 @@ func (db *Store) GetServerPermissions(ctx context.Context) ([]ServerPermission, 
 			flags = "z"
 		}
 
-		perms = append(perms, ServerPermission{
+		perms = append(perms, model.ServerPermission{
 			SteamID:         steamid.SID64ToSID(steamid.New(sid)),
 			PermissionLevel: perm,
 			Flags:           flags,
@@ -158,7 +83,7 @@ type ServerQueryFilter struct {
 	IncludeDisabled bool `json:"include_disabled"`
 }
 
-func (db *Store) GetServers(ctx context.Context, filter ServerQueryFilter) ([]Server, int64, error) {
+func (db *Store) GetServers(ctx context.Context, filter ServerQueryFilter) ([]model.Server, int64, error) {
 	builder := db.sb.
 		Select("s.server_id", "s.short_name", "s.name", "s.address", "s.port", "s.rcon", "s.password",
 			"s.token_created_on", "s.created_on", "s.updated_on", "s.reserved_slots", "s.is_enabled", "s.region", "s.cc",
@@ -187,15 +112,15 @@ func (db *Store) GetServers(ctx context.Context, filter ServerQueryFilter) ([]Se
 
 	rows, errQueryExec := db.QueryBuilder(ctx, builder)
 	if errQueryExec != nil {
-		return []Server{}, 0, Err(errQueryExec)
+		return []model.Server{}, 0, Err(errQueryExec)
 	}
 
 	defer rows.Close()
 
-	var servers []Server
+	var servers []model.Server
 
 	for rows.Next() {
-		var server Server
+		var server model.Server
 		if errScan := rows.
 			Scan(&server.ServerID, &server.ShortName, &server.Name, &server.Address, &server.Port, &server.RCON,
 				&server.Password, &server.TokenCreatedOn, &server.CreatedOn, &server.UpdatedOn, &server.ReservedSlots,
@@ -222,7 +147,7 @@ func (db *Store) GetServers(ctx context.Context, filter ServerQueryFilter) ([]Se
 	return servers, count, nil
 }
 
-func (db *Store) GetServerByName(ctx context.Context, serverName string, server *Server, disabledOk bool, deletedOk bool) error {
+func (db *Store) GetServerByName(ctx context.Context, serverName string, server *model.Server, disabledOk bool, deletedOk bool) error {
 	and := sq.And{sq.Eq{"short_name": serverName}}
 	if !disabledOk {
 		and = append(and, sq.Eq{"is_enabled": true})
@@ -254,7 +179,7 @@ func (db *Store) GetServerByName(ctx context.Context, serverName string, server 
 		&server.Deleted, &server.LogSecret, &server.EnableStats))
 }
 
-func (db *Store) GetServerByPassword(ctx context.Context, serverPassword string, server *Server, disabledOk bool, deletedOk bool) error {
+func (db *Store) GetServerByPassword(ctx context.Context, serverPassword string, server *model.Server, disabledOk bool, deletedOk bool) error {
 	and := sq.And{sq.Eq{"password": serverPassword}}
 	if !disabledOk {
 		and = append(and, sq.Eq{"is_enabled": true})
@@ -281,7 +206,7 @@ func (db *Store) GetServerByPassword(ctx context.Context, serverPassword string,
 }
 
 // SaveServer updates or creates the server data in the database.
-func (db *Store) SaveServer(ctx context.Context, server *Server) error {
+func (db *Store) SaveServer(ctx context.Context, server *model.Server) error {
 	server.UpdatedOn = time.Now()
 	if server.ServerID > 0 {
 		return db.updateServer(ctx, server)
@@ -292,7 +217,7 @@ func (db *Store) SaveServer(ctx context.Context, server *Server) error {
 	return db.insertServer(ctx, server)
 }
 
-func (db *Store) insertServer(ctx context.Context, server *Server) error {
+func (db *Store) insertServer(ctx context.Context, server *model.Server) error {
 	const query = `
 		INSERT INTO server (
 		    short_name, name, address, port, rcon, token_created_on, 
@@ -313,7 +238,7 @@ func (db *Store) insertServer(ctx context.Context, server *Server) error {
 	return nil
 }
 
-func (db *Store) updateServer(ctx context.Context, server *Server) error {
+func (db *Store) updateServer(ctx context.Context, server *model.Server) error {
 	server.UpdatedOn = time.Now()
 
 	return db.ExecUpdateBuilder(ctx, db.sb.
